misc: Don't call disconnectAll on token proxy objects
The disconnectAll method is a method available to javascript]
objects that want gobject-ish signal handling features.

It is not a method available to GObjects come from introspection.

"token" objects in the smartcard manager code are just plain old
GDBusProxy objects and so don't have disconnectAll methods.

This commit changes the code to use a direct disconnect call for
the one pertinent signal connection instead.